feat: Add getContent function to retrieve script content
This commit adds a new function, getContent, to the addon module in
order to retrieve script content from a specified file and script ID.
The getContent function accepts a file path and script ID as parameters
and returns the corresponding script content as a string. This
enhancement aims to provide a convenient way to access and manipulate
script content within the application. The addition of this function
expands the capabilities of the addon module, making it easier for
developers to work with script data.